feat(message): Hexagonal Architecture 전환 (#33)
- Message 서브도메인 헥사고날 구조 적용
- Conversation 서브도메인 헥사고날 구조 적용
- SSE 서브도메인 헥사고날 구조 적용
- Entity를 adapter/out/persistence/entity로 이동
- Repository를 JpaRepository로 리네이밍
- Controller를 adapter/in/web으로 이동
- DTOs를 adapter/in/web/dto로 이동
- Service를 application/service로 이동
